requirements.txt
amqp==5.2.0
appdirs==1.4.4
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
beautifulsoup4==4.12.3
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
channels==4.1.0
channels-redis==4.2.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
coverage==7.6.9
cron-descriptor==1.4.5
cryptography==43.0.3
daphne==4.1.2
diff_cover==9.2.0
Django==5.1.1
django-celery-beat==2.7.0
django-js-asset==2.2.0
django-mptt==0.16.0
django-recaptcha==4.0.0
django-timezone-field==7.0
django-tinymce==4.1.0
flower==2.0.1
h11==0.14.0
humanize==4.11.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
iniconfig==2.0.0
isort==5.13.2
Jinja2==3.1.4
kombu==5.4.2
lxml==5.3.0
MarkupSafe==2.1.5
msgpack==1.1.0
outcome==1.3.0.post0
packaging==24.1
pathspec==0.12.1
pillow==10.4.0
pluggy==1.5.0
prometheus_client==0.21.1
prompt_toolkit==3.0.48
psycopg2==2.9.9
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
Pygments==2.18.0
pyOpenSSL==24.2.1
PySocks==1.7.1
pytest==8.3.3
pytest-django==4.9.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-memcached==1.62
pytz==2024.2
PyYAML==6.0.2
redis==5.1.1
regex==2024.9.11
requests==2.32.3
ruff==0.5.7
selenium==4.25.0
service-identity==24.2.0
setuptools==75.3.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sqlfluff==3.1.1
sqlparse==0.5.1
tblib==3.0.0
tornado==6.4.2
tqdm==4.66.5
trio==0.26.2
trio-websocket==0.11.1
Twisted==24.10.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
websocket-client==1.8.0
wsproto==1.2.0
zope.interface==7.1.1
